fix: DLQ와 SSE 요청에서 DTO 형식이 다른 오류 수정
- SSE 요청을 보낼 때 DTO와 DLQ 처리할 떄 DTO가 다른 문제 수정
